simplify logic for usability
This updates functionality to be more intuitive. Now, regardless of if you’re editing a pm or creating a new one, the app will fetch the entire list of signals, sorted by closest to your location. If this is a new PM, we will populate the signal field automatically with the nearest signal, but the user can modify it.

We've done away with the mess of conditionally searching for nearby signals, then subsequently searching for all signals if nothing nearby is found. It's also no longer to use a "Reset Signal Selection" button, so that logic is also removed. I also cleaned up the `SETINTERVAL()` logic to reduce redundancy